All tracks into the json
